ayyy, lmao. It works.
Changed String to char[]; it might be a slightly larger buffer size, but it's worth saving CPU instructions ( or something :^) )
Changed around the transferFrom(Peer) method on a lot of Transmittable objects to use a new peer's IN_BUFFER rather than reading from the channel again; we don't need to cause a ReadPendingException.
Fixed some slight "bug" in HashUtils that caused it to run very slowly. Remember to read stuff in chunks and not 1 byte at a time, kids.

+++ src/main/java/ga/nurupeaches/katou/chunk/FileChunk.java
@@ -10,6 +10,9 @@ import java.nio.channels.FileLock;
 import java.util.logging.Level;
 import java.util.zip.CRC32;
 
+/**
+ * Represents a chunk of a file on disk.
+ */
 public class FileChunk extends RepresentableChunk {
 
     private AsynchronousFileChannel fileChannel;
@@ -29,17 +32,17 @@ public class FileChunk extends RepresentableChunk {
 
     @Override
     public void transferFrom(Peer peer) throws IOException{
-
+        // TODO: impl
     }
 
     @Override
     public void transferTo(Peer peer) throws IOException{
-
+        // TODO: impl
     }
 
     @Override
     public boolean validate(long crc32){
-        ByteBuffer buffer = ByteBuffer.allocate(getSize());
+        ByteBuffer buffer = ByteBuffer.allocate((int)getSize());
 
         try {
             FileLock lock = fileChannel.tryLock(index, getSize(), false);
